Add files via upload
This project considers data generated based on a linear separator. I discuss the relationship of the cut that defines the classes of the data with the average of many random linear cuts that perfectly classify the data, which comes very close to the maximum marginal classifier. I then consider linear classifiers that misclassify one or more of the data points, and how the pool of viable cuts relates to the number of misclassified points, and perhaps if there is a way to calculate a weighted average of the random cuts that takes into account the number of misclassified points.